fix: Move tests out of folder
